scanner/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
urllib3>=2.0.0
python-whois>=0.9.0
dnspython>=2.4.0
pypdf>=4.0.0
Pillow>=10.0.0
